Fixed bug in refresh task. It now works.

+++ maptak/app/src/main/java/com/droptak/android/tasks/GetUsersMapsTask.java
@@ -3,6 +3,7 @@ package com.droptak.android.tasks;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.os.AsyncTask;
+import android.util.Log;
 import android.widget.Toast;
 
 import com.droptak.android.activities.MainActivity;
@@ -44,7 +45,7 @@ public class GetUsersMapsTask extends AsyncTask<Void, Void, Void> {
         SharedPreferences prefs = c.getSharedPreferences(MainActivity.SHARED_PREFS_NAME, 0);
         String userID = prefs.getString(MainActivity.PREF_USER_MAPTAK_TOKEN, "");
         if (userID.equals("")) {
-            Toast.makeText(c, "Error downloading maps. Not signed in.", Toast.LENGTH_SHORT).show();
+            Log.d(MainActivity.LOG_TAG, "Error refreshing local DB. User not signed in");
             return null;
         }
         String url = BASE_URL + userID + "/maps";
@@ -86,15 +87,20 @@ public class GetUsersMapsTask extends AsyncTask<Void, Void, Void> {
             // Get the old version of the map in the event data is updated
             MapObject oldMap = db.getMap(map.getID());
 
-            // Delete the map from the database
-            db.deleteMap(oldMap.getID());
+            // If the old map exists in the database
+            if (oldMap != null) {
+                // Delete it
+                db.deleteMap(oldMap.getID());
+            }
 
             // And re-add it
             db.addMap(map);
 
             // Delete its admins and re-add them
-            for (User u : oldMap.getManagers()) {
-                db.deleteAdmin(u, oldMap.getID());
+            if (oldMap != null) {
+                for (User u : oldMap.getManagers()) {
+                    db.deleteAdmin(u, oldMap.getID());
+                }
             }
 
             // Re-add them
@@ -103,16 +109,18 @@ public class GetUsersMapsTask extends AsyncTask<Void, Void, Void> {
             }
 
             // Delete its taks
-            for (TakObject t : oldMap.getTaks()) {
-                db.deleteTak(t.getID());
+            if (oldMap != null) {
+                for (TakObject t : oldMap.getTaks()) {
+                    db.deleteTak(t.getID());
 
-                // Delete its metadata
-                for (TakMetadata m : t.getMeta().values()) {
-                    db.deleteTakMetadata(m);
+                    // Delete its metadata
+                    for (TakMetadata m : t.getMeta().values()) {
+                        db.deleteTakMetadata(m);
+                    }
                 }
             }
 
-            // Re-add htem
+            // Re-add them
             for (TakObject t : map.getTaks()) {
                 db.addTak(t, map.getID());
 
@@ -123,8 +131,6 @@ public class GetUsersMapsTask extends AsyncTask<Void, Void, Void> {
             }
 
         }
-
-
 
         return null;
     }
